Change SeekableStream::len to u64
Aligns with file size and will not overflow on 32-bit architectures.

+++ sdk/core/typespec_client_core/src/http/request/mod.rs
@@ -37,9 +37,9 @@ pub enum Body {
 
 impl Body {
     /// Returns the length of the body in bytes.
-    pub fn len(&self) -> usize {
+    pub fn len(&self) -> u64 {
         match self {
-            Body::Bytes(bytes) => bytes.len(),
+            Body::Bytes(bytes) => bytes.len() as u64,
             Body::SeekableStream(stream) => stream.len(),
         }
     }
